# docs/architecture/api-spec.yaml
openapi: 3.1.0
info:
  title: Enginex AI API
  version: 1.0.0
  description: Auto-generated from the FastAPI app (see app/scripts/export_openapi.py) — this is the same
    schema served live at /openapi.json. WebSocket endpoints (/ws/files/{file_id}) aren't representable
    in OpenAPI 3.0 and are documented separately in docs/architecture/complete-architecture.md and services/backend/app/websockets/manager.py.
paths:
...
  /api/v1/cad/sketches:
    post:
      tags:
      - CAD
      summary: Create Sketch
      operationId: create_sketch_api_v1_cad_sketches_post
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateSketchRequest'
        required: true
      responses:
        '201':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CADObjectOut'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
      security:
      - HTTPBearer: []
  /api/v1/cad/sketches/{sketch_id}:
    get:
      tags:
      - CAD
      summary: Get Sketch
      operationId: get_sketch_api_v1_cad_sketches__sketch_id__get
      security:
      - HTTPBearer: []
      parameters:
      - name: sketch_id
        in: path
        required: true
        schema:
          type: string
          format: uuid
          title: Sketch Id
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CADObjectOut'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
    put:
      tags:
      - CAD
      summary: Update Sketch
      operationId: update_sketch_api_v1_cad_sketches__sketch_id__put
      security:
      - HTTPBearer: []
      parameters:
      - name: sketch_id
        in: path
        required: true
        schema:
          type: string
          format: uuid
          title: Sketch Id
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateSketchRequest'
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CADObjectOut'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
    delete:
      tags:
      - CAD
      summary: Delete Sketch
      operationId: delete_sketch_api_v1_cad_sketches__sketch_id__delete
      security:
      - HTTPBearer: []
      parameters:
      - name: sketch_id
        in: path
        required: true
        schema:
          type: string
          format: uuid
          title: Sketch Id
      responses:
        '204':
          description: Successful Response
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/v1/cad/sketches/{sketch_id}/constraints:
    post:
      tags:
      - CAD
      summary: Add Constraint
      operationId: add_constraint_api_v1_cad_sketches__sketch_id__constraints_post
      security:
      - HTTPBearer: []
      parameters:
      - name: sketch_id
        in: path
        required: true
        schema:
          type: string
          format: uuid
          title: Sketch Id
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema: {}
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/v1/cad/bodies:
    get:
      tags:
      - CAD
      summary: List Bodies
      operationId: list_bodies_api_v1_cad_bodies_get
      security:
      - HTTPBearer: []
      parameters:
      - name: file_id
        in: query
        required: true
        schema:
          type: string
          format: uuid
          title: File Id
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/CADObjectOut'
                title: Response List Bodies Api V1 Cad Bodies Get
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/v1/cad/bodies/{body_id}/extrude:
    post:
      tags:
      - CAD
      summary: Extrude
      operationId: extrude_api_v1_cad_bodies__body_id__extrude_post
      security:
      - HTTPBearer: []
      parameters:
      - name: body_id
        in: path
        required: true
        schema:
          type: string
          format: uuid
          title: Body Id
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema: {}
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/v1/cad/bodies/{body_id}/revolve:
    post:
      tags:
      - CAD
      summary: Revolve
      operationId: revolve_api_v1_cad_bodies__body_id__revolve_post
      security:
      - HTTPBearer: []
      parameters:
      - name: body_id
        in: path
        required: true
        schema:
          type: string
          format: uuid
          title: Body Id
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema: {}
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/v1/cad/bodies/{body_id}/fillet:
    post:
      tags:
      - CAD
      summary: Fillet
      operationId: fillet_api_v1_cad_bodies__body_id__fillet_post
      security:
      - HTTPBearer: []
      parameters:
      - name: body_id
        in: path
        required: true
        schema:
          type: string
          format: uuid
          title: Body Id
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema: {}
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/v1/cad/bodies/{body_id}/chamfer:
    post:
      tags:
      - CAD
      summary: Chamfer
      operationId: chamfer_api_v1_cad_bodies__body_id__chamfer_post
      security:
      - HTTPBearer: []
      parameters:
      - name: body_id
        in: path
        required: true
        schema:
          type: string
          format: uuid
          title: Body Id
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema: {}
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/v1/cad/export/step/{body_id}:
    get:
      tags:
      - CAD
      summary: Export Step
      operationId: export_step_api_v1_cad_export_step__body_id__get
      security:
      - HTTPBearer: []
      parameters:
      - name: body_id
        in: path
        required: true
        schema:
          type: string
          format: uuid
          title: Body Id
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema: {}
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/v1/cad/export/stl/{body_id}:
    get:
      tags:
      - CAD
      summary: Export Stl
      operationId: export_stl_api_v1_cad_export_stl__body_id__get
      security:
      - HTTPBearer: []
      parameters:
      - name: body_id
        in: path
        required: true
        schema:
          type: string
          format: uuid
          title: Body Id
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema: {}
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
...
  /api/v1/ai/chats:
    post:
      tags:
      - AI
      summary: Create Chat
      operationId: create_chat_api_v1_ai_chats_post
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateChatRequest'
        required: true
      responses:
        '201':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ChatOut'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
      security:
      - HTTPBearer: []
  /api/v1/ai/chats/{chat_id}:
    get:
      tags:
      - AI
      summary: Get Chat
      operationId: get_chat_api_v1_ai_chats__chat_id__get
      security:
      - HTTPBearer: []
      parameters:
      - name: chat_id
        in: path
        required: true
        schema:
          type: string
          format: uuid
          title: Chat Id
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ChatOut'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
    delete:
      tags:
      - AI
      summary: Delete Chat
      operationId: delete_chat_api_v1_ai_chats__chat_id__delete
      security:
      - HTTPBearer: []
      parameters:
      - name: chat_id
        in: path
        required: true
        schema:
          type: string
          format: uuid
          title: Chat Id
      responses:
        '204':
          description: Successful Response
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/v1/ai/chats/{chat_id}/messages:
    post:
      tags:
      - AI
      summary: Post Message
      operationId: post_message_api_v1_ai_chats__chat_id__messages_post
      security:
      - HTTPBearer: []
      parameters:
      - name: chat_id
        in: path
        required: true
        schema:
          type: string
          format: uuid
          title: Chat Id
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateMessageRequest'
      responses:
        '201':
          description: Successful Response
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/MessageOut'
                title: Response Post Message Api V1 Ai Chats  Chat Id  Messages Post
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
    get:
      tags:
      - AI
      summary: List Messages
      operationId: list_messages_api_v1_ai_chats__chat_id__messages_get
      security:
      - HTTPBearer: []
      parameters:
      - name: chat_id
        in: path
        required: true
        schema:
          type: string
          format: uuid
          title: Chat Id
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/MessageOut'
                title: Response List Messages Api V1 Ai Chats  Chat Id  Messages Get
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/v1/ai/agents:
    get:
      tags:
      - AI
      summary: List Agents
      operationId: list_agents_api_v1_ai_agents_get
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                items:
                  $ref: '#/components/schemas/AgentOut'
                type: array
                title: Response List Agents Api V1 Ai Agents Get
      security:
      - HTTPBearer: []
  /api/v1/ai/providers:
    get:
      tags:
      - AI
      summary: List Providers
      operationId: list_providers_api_v1_ai_providers_get
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                items:
                  $ref: '#/components/schemas/ProviderOut'
                type: array
                title: Response List Providers Api V1 Ai Providers Get
      security:
      - HTTPBearer: []
  /api/v1/ai/providers/configure:
    post:
      tags:
      - AI
      summary: Configure Provider
      operationId: configure_provider_api_v1_ai_providers_configure_post
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ConfigureProviderRequest'
        required: true
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema: {}
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
      security:
      - HTTPBearer: []
...
components:
  schemas:
...
    AgentOut:
      properties:
        id:
          type: string
          format: uuid
          title: Id
        name:
          type: string
          title: Name
        role:
          type: string
          title: Role
        prompt:
          type: string
          title: Prompt
        tools:
          items: {}
          type: array
          title: Tools
        memory_enabled:
          type: boolean
          title: Memory Enabled
      type: object
      required:
      - id
      - name
      - role
      - prompt
      - tools
      - memory_enabled
      title: AgentOut
...
    CADObjectOut:
      properties:
        id:
          type: string
          format: uuid
          title: Id
        file_id:
          type: string
          format: uuid
          title: File Id
        object_type:
          type: string
          title: Object Type
        name:
          type: string
          title: Name
        data:
          additionalProperties: true
          type: object
          title: Data
        parent_id:
          anyOf:
          - type: string
            format: uuid
          - type: 'null'
          title: Parent Id
        version_number:
          type: integer
          title: Version Number
        created_at:
          type: string
          format: date-time
          title: Created At
        updated_at:
          type: string
          format: date-time
          title: Updated At
      type: object
      required:
      - id
      - file_id
      - object_type
      - name
      - data
      - parent_id
      - version_number
      - created_at
      - updated_at
      title: CADObjectOut
    ChatOut:
      properties:
        id:
          type: string
          format: uuid
          title: Id
        project_id:
          anyOf:
          - type: string
            format: uuid
          - type: 'null'
          title: Project Id
        user_id:
          type: string
          format: uuid
          title: User Id
        title:
          type: string
          title: Title
        created_at:
          type: string
          format: date-time
          title: Created At
        updated_at:
          type: string
          format: date-time
          title: Updated At
      type: object
      required:
      - id
      - project_id
      - user_id
      - title
      - created_at
      - updated_at
      title: ChatOut
    ConfigureProviderRequest:
      properties:
        provider:
          type: string
          title: Provider
        api_key:
          type: string
          title: Api Key
      type: object
      required:
      - provider
      - api_key
      title: ConfigureProviderRequest
...
    CreateChatRequest:
      properties:
        project_id:
          anyOf:
          - type: string
            format: uuid
          - type: 'null'
          title: Project Id
        title:
          type: string
          maxLength: 255
          minLength: 1
          title: Title
          default: New chat
      type: object
      title: CreateChatRequest
...
    CreateMessageRequest:
      properties:
        content:
          type: string
          minLength: 1
          title: Content
      type: object
      required:
      - content
      title: CreateMessageRequest
...
    CreateSketchRequest:
      properties:
        file_id:
          type: string
          format: uuid
          title: File Id
        name:
          type: string
          maxLength: 255
          minLength: 1
          title: Name
        data:
          additionalProperties: true
          type: object
          title: Data
      type: object
      required:
      - file_id
      - name
      title: CreateSketchRequest
...
    HTTPValidationError:
      properties:
        detail:
          items:
            $ref: '#/components/schemas/ValidationError'
          type: array
          title: Detail
      type: object
      title: HTTPValidationError
...
    MessageOut:
      properties:
        id:
          type: string
          format: uuid
          title: Id
        chat_id:
          type: string
          format: uuid
          title: Chat Id
        role:
          type: string
          title: Role
        content:
          type: string
          title: Content
        tool_calls:
          items: {}
          type: array
          title: Tool Calls
        model_used:
          anyOf:
          - type: string
          - type: 'null'
          title: Model Used
        tokens_used:
          type: integer
          title: Tokens Used
        created_at:
          type: string
          format: date-time
          title: Created At
      type: object
      required:
      - id
      - chat_id
      - role
      - content
      - tool_calls
      - model_used
      - tokens_used
      - created_at
      title: MessageOut
...
    ProviderOut:
      properties:
        name:
          type: string
          title: Name
        configured:
          type: boolean
          title: Configured
      type: object
      required:
      - name
      - configured
      title: ProviderOut
...
    UpdateSketchRequest:
      properties:
        name:
          anyOf:
          - type: string
            maxLength: 255
            minLength: 1
          - type: 'null'
          title: Name
        data:
          anyOf:
          - additionalProperties: true
            type: object
          - type: 'null'
          title: Data
      type: object
      title: UpdateSketchRequest
...
    ValidationError:
      properties:
        loc:
          items:
            anyOf:
            - type: string
            - type: integer
          type: array
          title: Location
        msg:
          type: string
          title: Message
        type:
          type: string
          title: Error Type
        input:
          title: Input
        ctx:
          type: object
          title: Context
      type: object
      required:
      - loc
      - msg
      - type
      title: ValidationError
...
  securitySchemes:
    HTTPBearer:
      type: http
      scheme: bearer
servers:
- url: http://localhost:8000
  description: Local development
- url: https://api.enginex.ai
  description: Production (placeholder)